Make installation more portable
In this patch I make two significant changes:

First, I use the installed dune `dt` binary to calculate the
dependencyTree of a Node.js module. This change is accompanied
by introduction of the `Bos` library to call subprocesses in
a more declarative manner.

The big change above assumes that the `dt` was installed
correctly in the change introduced in 3516167. This commit
requires `npm` to be configured correctly. Here I add the
`conf-npm` package to try and do this automatically, but
some systems may require more manual configuration.
Regardless, if compilation succeeds then everything will be ok.

Lastly, I create a `share/config` site for the `ast_gen`
package to install the default `config.json` file.
This extension makes the installation more portable as sites
are managed by dune, and are configured in the `Auxiliary.Share`
module.